requirements.txt
matplotlib==3.5.1
matplotlib-inline==0.1.3
numpy==1.20.3
pandas==1.3.5
pymc3==3.11.2
scipy==1.7.3
seaborn==0.11.2
Theano-PyMC==1.1.2
ipykernel==6.6.1